Fix compilation with current swig and numpy
With thanks to Fei Tan for reporting the issue.

In the current versions of SWIG and python, the numpy includes directory
containing numpy/arrayobject.h which is required by SWIG during
compilation is not added to the include path of the gcc invocation. This
patch adds the appropriate path to the include_dirs variable for the
msacounts extension.